requirements.txt
aiomysql==0.0.22
anyio==3.5.0
asgiref==3.5.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
databases==0.5.5
ecdsa==0.17.0
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
mysqlclient==2.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyMySQL==0.9.3
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
watchgod==0.7
websockets==10.2
